+++ src/converse-profile.js
@@ -46,7 +46,7 @@ converse.plugins.add('converse-profile', {
             },
 
             initialize () {
-                this.model.on('change', this.render, this);
+                this.listenTo(this.model, 'change', this.render);
                 _converse.BootstrapModal.prototype.initialize.apply(this, arguments);
                 /**
                  * Triggered when the _converse.ProfileModal has been created and initialized.
@@ -234,8 +234,8 @@ converse.plugins.add('converse-profile', {
             },
 
             initialize () {
-                this.model.on("change", this.render, this);
-                this.model.vcard.on("change", this.render, this);
+                this.listenTo(this.model, "change", this.render);
+                this.listenTo(this.model.vcard, "change", this.render);
             },
 
             toHTML () {